requirements.txt
appdirs==1.4.4
asgiref==3.3.4
attrs==21.2.0
cached-property==1.5.2
certifi==2020.12.5
chardet==4.0.0
defusedxml==0.7.1
Django==3.2.3
django-filter==2.4.0
django-redis==4.12.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
idna==2.10
importlib-metadata==4.0.1
isodate==0.6.0
lxml==4.6.3
Markdown==3.3.4
Pillow==8.2.0
psycopg2==2.8.6
PyJWT==2.1.0
python-decouple==3.4
pytz==2021.1
redis==3.5.3
requests==2.25.1
requests-file==1.5.1
requests-toolbelt==0.9.1
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
urllib3==1.26.4
zeep==4.0.0
zipp==3.4.1
